+++ src/app/pen-test/web-app-pen-test/[pentestIdName]/components/PentestTabs.tsx
@@ -4,6 +4,7 @@ import React, { useEffect, useState } from 'react';
 import { Card } from '@tremor/react';
 import { usePathname, useRouter } from 'next/navigation';
 import { Loader2 } from 'lucide-react';
+import GlobalLoader from '@/components/GlobalLoader';
 
 export default function PentestTabs({ pentestIdName }: { pentestIdName: string }) {
     const [pentestId, pentestNameInUrl] = pentestIdName.split("_");
@@ -37,11 +38,6 @@ export default function PentestTabs({ pentestIdName }: { pentestIdName: string }
         <Card className="bg-tremor-background-muted p-0 dark:bg-dark-tremor-background-muted">
             <nav className="flex px-4 gap-3">
                 {tabData.map((eachTab, index) => {
-                    // const isActive = pathname.includes(eachTab.pentestState.toLowerCase().replace(/\s+/g, '-'));
-                    // const isActive = pathname.startsWith(eachTab.href);
-                    // const updatedPathName = pathname.replace(/-/g, '');
-                    // const eachTabNameToMatchRoute = eachTab.pentestState.toLowerCase().replace(/\s+/g, '');
-                    // const isActive = updatedPathName.includes(eachTabNameToMatchRoute);
                     const eachTabUpdated = eachTab.pentestState.split(" ").map((part, index) => index === 0 ? part.toLowerCase() : part).join("");
                     const pathNameArray = pathname.split("/");
                     let isActive = false;
@@ -56,7 +52,11 @@ export default function PentestTabs({ pentestIdName }: { pentestIdName: string }
                         }
                     }
 
-                    const isLoading = loadingTab === eachTab.pentestState;
+                    // const isLoading = loadingTab === eachTab.pentestState;
+
+                    if (loadingTab === eachTab.pentestState) {
+                        return <GlobalLoader key={eachTab.pentestState} />
+                    }
 
                     return (
                         <button
@@ -65,11 +65,12 @@ export default function PentestTabs({ pentestIdName }: { pentestIdName: string }
                             className={`p-2.5 font-medium flex gap-2 items-center ${isActive
                                 ? 'border-b-2 border-blue-500 text-blue-600'
                                 : 'hover:border-gray-300'
-                                } ${isLoading ? 'pointer-events-none' : ''}`}
+                                }`}
+                        // ${isLoading ? 'pointer-events-none' : ''}`}
                         >
                             <span>{index + 1}.</span>
                             {eachTab.pentestState}
-                            {isLoading && <span className='tab-loader'></span>}
+                            {/* {isLoading && <span className='tab-loader'></span>} */}
                         </button>
                     );
                 })}
