+++ src/app/pen-test/web-app-pen-test/[pentestIdName]/reconnaissance/active-reconnaissance/components/ActiveReconnaissanceSidebar.tsx
@@ -3,6 +3,7 @@
 import React, { useEffect, useState } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
 import { Cpu, EthernetPort, LayoutDashboard, Network, Loader2 } from 'lucide-react';
+import GlobalLoader from '@/components/GlobalLoader';
 
 export default function ActiveReconnaissanceSidebar({ pentestIdName }: { pentestIdName: string }) {
   const [pentestId, pentestNameInUrl] = pentestIdName.split("_");
@@ -60,7 +61,11 @@ export default function ActiveReconnaissanceSidebar({ pentestIdName }: { pentest
       <div className="space-y-1">
         {tabData.map((eachTab) => {
           const isActive = pathname.includes(eachTab.tabName.toLowerCase().replace(/\s+/g, '-'));
-          const isLoading = loadingTab === eachTab.tabName;
+          // const isLoading = loadingTab === eachTab.tabName;
+
+          if (loadingTab === eachTab.tabName) {
+            return <GlobalLoader key={eachTab.tabName} />
+          }
 
           return (
             <button
@@ -68,12 +73,12 @@ export default function ActiveReconnaissanceSidebar({ pentestIdName }: { pentest
               onClick={() => handleTabClick(eachTab.href, eachTab.tabName)}
               className={`flex w-full items-center gap-x-2.5 px-3 py-2 text-sm rounded-md transition-all truncate
                 ${isActive ? 'bg-primary text-white' :
-                  'text-gray-700 dark:text-gray-400 hover:bg-gray-100 hover:text-gray-900 hover:dark:text-gray-50 hover:dark:bg-gray-800'}
-                ${isLoading ? 'pointer-events-none' : ''}`}
+                  'text-gray-700 dark:text-gray-400 hover:bg-gray-100 hover:text-gray-900 hover:dark:text-gray-50 hover:dark:bg-gray-800'}`}
+            // ${isLoading ? 'pointer-events-none' : ''}`}
             >
               <eachTab.Icon className="-ml-1 size-4" aria-hidden="true" />
               {eachTab.tabName}
-              {isLoading && <span className='tab-loader'></span>}
+              {/* {isLoading && <span className='tab-loader'></span>} */}
             </button>
           );
         })}
