+++ src/app/pen-test/web-app-pen-test/[pentestIdName]/reconnaissance/components/ReconnaissanceTabs.tsx
@@ -4,6 +4,7 @@ import React, { useEffect, useState } from 'react';
 import { Card } from '@tremor/react';
 import { usePathname, useRouter } from 'next/navigation';
 import { ScanSearch, TextSearch, Loader2 } from 'lucide-react';
+import GlobalLoader from '@/components/GlobalLoader';
 
 export default function ReconnaissanceTabs({ pentestIdName }: { pentestIdName: string }) {
     const [pentestId, pentestNameInUrl] = pentestIdName.split("_");
@@ -48,7 +49,11 @@ export default function ReconnaissanceTabs({ pentestIdName }: { pentestIdName: s
             <nav className="flex gap-3">
                 {tabData.map((eachTab) => {
                     const isActive = pathname.includes(eachTab.tabName.toLowerCase().replace(/\s+/g, '-'));
-                    const isLoading = loadingTab === eachTab.tabName;
+                    // const isLoading = loadingTab === eachTab.tabName;
+
+                    if (loadingTab === eachTab.tabName) {
+                        return <GlobalLoader key={eachTab.tabName} />
+                    }
 
                     return (
                         <button
@@ -57,11 +62,12 @@ export default function ReconnaissanceTabs({ pentestIdName }: { pentestIdName: s
                             className={`p-2.5 font-medium flex gap-1.5 items-center ${isActive
                                 ? 'border-b-2 border-blue-500 text-blue-600'
                                 : 'hover:border-gray-300'
-                                } ${isLoading ? 'pointer-events-none' : ''}`}
+                                }`}
+                        // ${isLoading ? 'pointer-events-none' : ''}`}
                         >
                             <eachTab.Icon className="-ml-1 size-4" aria-hidden="true" />
                             {eachTab.tabName}
-                            {isLoading && <span className='tab-loader'></span>}
+                            {/* {isLoading && <span className='tab-loader'></span>} */}
                         </button>
                     );
                 })}
