+++ src/app/pen-test/web-app-pen-test/[pentestIdName]/reconnaissance/loading.tsx
@@ -1,3 +1,5 @@
+import GlobalLoader from "@/components/GlobalLoader";
+
 export default function Loading() {
-    return "Loading...";
+    return <GlobalLoader />;
 }
